+++ CHANGES.rst
@@ -21,11 +21,13 @@ Project `milestones <https://github.com/BCDA-APS/apstools/milestones>`_
 describe future plans.
 
 ..
-   .. 
-      future: 1 . 8 . 0
+   ..
+      1.8.0
       ******
 
-      Add async devices.  Release expected by 2026-10-10.
+      Release expected by 2026-10-10.
+
+      Add async devices.
 
    1.7.11
    ******